Finished translating text_wrap.py into test.el
I realised that in line 80 of test.el, our method of determining whether to add text or add whitespace is rendered overcomplicated. Because `word_reduced` is always going to be (1 1 1 ... 1) up to length of the word. This also happens in the Python code though we didn't notice. Originally i think it was meant to be (1 1 1 ... 1 0) so there was some use to it. Anyway the xor comparision then boils down to how many of `look_ahead` is 1's, in which case we can just do the original (/ (apply '+ look_ahead) (length look_ahead)). We'll keep this for now in case we do need the extra complexity but otherwise we should simplify the code.

TODO: pull out the necessary functions out of test.el into its own self contained progn script